Consolidate ingestion and linking code
Ingestion and linking code are a part of one step
and therefore should belong to the same submodule.
Removed defunct R code.
