--- PIR_Pipeline/src/pir_pipeline/pir_ingestion/tests/test_ingest_data.R
@@ -1,14 +1,0 @@
-test_that(
-  "Input file types are correct",
-  expect_true(
-    all(
-      map_lgl(
-        wb_list,
-        ~ grepl(".xlsx?", .x) 
-      )
-    ) == TRUE
-  )
-)
-test_that(
-  ""
-)
