--- PIR_Pipeline/src/pir_pipeline/pir_ingestion/utils/addUnmatched.R
@@ -1,20 +1,0 @@
-################################################################################
-## Written by: Reggie Gilliard
-## Date: 11/10/2023
-## Description: Add unmatched variables to a data frame
-################################################################################
-
-
-#' Add the variable `unmatched` to a data frame
-#' 
-#' @param data Data frame
-#' @returns A data frame containing the variable `unmatched`.
-#' @examples
-#' addUnmatched(mtcars)
-
-addUnmatched <- function(data) {
-  if ("unmatched" %notin% names(data)) {
-    data <- dplyr::mutate(data, unmatched = NA_character_)
-  }
-  return(data)
-}
