--- PIR_Pipeline/src/pir_pipeline/pir_question_links/utils/genUQID.R
@@ -1,36 +1,0 @@
-################################################################################
-## Written by: Reggie Gilliard
-## Date: 01/02/2023
-## Description: Script to generate a unique question ID.
-################################################################################
-
-
-#' Generate a unique question ID
-#' 
-#' `genUQID` checks for a `uqid` column and generates uqids for
-#' records without them.
-#' @param df A data frame.
-#' @returns The input data frame with an updated/generated `uqid` column. 
-
-genUQID <- function(df) {
-  
-  require(dplyr)
-  
-  df %>%
-    # Generate uqid if it does not exist. 
-    {
-      if (is.null(.$uqid)) {
-        mutate(., uqid = NA_character_)
-      } else {
-        .
-      }
-    } %>%
-    mutate(
-      uqid = case_when(
-        is.na(uqid) ~ uuid::UUIDgenerate(n = nrow(.)),
-        TRUE ~ uqid
-      )
-    ) %>%
-    assertr::assert(is_uniq, uqid) %>%
-    return()
-}
